+++ plustwoc/TwoChoiceStorage.h
@@ -30,28 +30,28 @@ private:
     vector<vector<vector<string>>> cuckoofilenames;
     prf_type nullKey;
     string fileAddressPrefix = "/tmp/";
-    int dataIndex;
-    vector<int> numberOfBins;
-    vector<int> sizeOfEachBin;
-    int KEY_VALUE_SIZE = (2 * AES_KEY_SIZE);
+    long dataIndex;
+    vector<long> numberOfBins;
+    vector<long> sizeOfEachBin;
+    long KEY_VALUE_SIZE = (2 * AES_KEY_SIZE);
     vector< vector<pair<prf_type, prf_type> > > data;
     //stxxl::VECTOR_GENERATOR< pair<prf_type, prf_type>, 4, 8, 1 * 1024 * 1024, stxxl::RC, stxxl::lru >::result** diskData;
 
 public:
-    int readBytes = 0;
-    int SeekG = 0;
-    TwoChoiceStorage(bool inMemory, int dataIndex, string fileAddressPrefix, bool profile);
+    long readBytes = 0;
+    long SeekG = 0;
+    TwoChoiceStorage(bool inMemory, long dataIndex, string fileAddressPrefix, bool profile);
     bool setup(bool overwrite);
-    void insertAll(int dataIndex, vector<vector< pair<prf_type, prf_type> > > ciphers);
-    void insertStash(int dataIndex, vector<prf_type> ciphers);
-    pair<prf_type, vector<prf_type>> insertCuckooHT(int index, int tableNum, int cuckooID, int hash, prf_type keyw, vector<prf_type> fileids);
-    vector<pair<prf_type, prf_type> > getAllData(int dataIndex);
-    vector<prf_type> getStash(int index); 
-    void insertCuckooStash(int index, int tableNum, vector<prf_type> ctCiphers);
-    vector<prf_type> getCuckooHT(int index); 
-    vector <prf_type> cuckooSearch(int index, int tableNum, int h[2]);
-    void clear(int index);
-    vector<prf_type> find(int index, prf_type mapKey, int cnt);
+    void insertAll(long dataIndex, vector<vector< pair<prf_type, prf_type> > > ciphers);
+    void insertStash(long dataIndex, vector<prf_type> ciphers);
+    pair<prf_type, vector<prf_type>> insertCuckooHT(long index, long tableNum, long cuckooID, long hash, prf_type keyw, vector<prf_type> fileids);
+    vector<pair<prf_type, prf_type> > getAllData(long dataIndex);
+    vector<prf_type> getStash(long index); 
+    void insertCuckooStash(long index, long tableNum, vector<prf_type> ctCiphers);
+    vector<prf_type> getCuckooHT(long index); 
+    vector <prf_type> cuckooSearch(long index, long tableNum, long h[2]);
+    void clear(long index);
+    vector<prf_type> find(long index, prf_type mapKey, long cnt);
     virtual ~TwoChoiceStorage();
 };
